requirements.txt
Django==1.8.4
dj-database-url==0.3.0
git+git://github.com/jezdez/django-configurations.git#egg=django-configurations
django-crispy-forms==1.5.1
django-dotenv==1.3.0
django-extensions==1.5.5
psycopg2==2.6.1
six==1.9.0
waitress==0.8.9
websockets==2.6
whitenoise==2.0.3
